Fix bull metrics collector to use proper redis client
We no longer have the `redisUrl` there available, instead we have a
`redis` property that may be created based on the passed `redis` or
`redis_url` options for instance, and now we use that correctly to setup
the bull queues.

This makes the fastify web + bull sample app work, using the same redis
URL between web & worker if a custom one is passed, and bull processes
jobs & reports queue stats to Judoscale correctly.