fix: Fix bull collector queue connection with Redis (#61)
This started as some minor copy fixes on the fastify + bull sample app, when I realized it wasn't fully working. It seems we were trying to pass a redis URL down to the the bull Queue, but we've since moved away from the URL in favor of the `redis` property.

This switches the url to passing `createClient`, which is how we've been doing it on our sample app. I couldn't find another way to pass the connection instance (which seems to be possible in bullmq).

There's also a few minor updates to the sample apps to allow a custom REDIS_URL to be passed for instance, and copy fixes on the fastify + bull sample as I had originally started.