Refactor express tests to go through the entire request life-cycle
Similar to fastify tests, but with express we're using `supertest` to
run the request and check for the response.

We're now also verifying the response is successful for both fastify &
express, besides our specific metrics expectations.

This makes express more of an integration / end-to-end test that runs
through the middleware, rather than the previous more unit-based testing
version.