Add test for elapsed time math
Took me some time to figure out a way to mock process hrtime.bigint
correctly, but it seems to be working. It does need the extra
`require('process')` which doesn't seem to be necessary for the
implementation to work, but the explicit require allows jest to mock it
so we can fix our return value to test with it.