Implement base utilization tracker module to track idle time
This brings over the utilization tracker implementation from Ruby/Python
to Node, essentially replicating the same logic and tests to verify
behavior of tracking idle time and converting that into utilization
percentage that we can then use to track as a metric soon.

I'm not sure about the need for a lock or similar concept though, it's
necessary in Ruby/Python because those are threaded systems / servers in
general, but that should not be the case with Node.