Always work with BigInt values when calculating total idle time
Since monotonic time returns BigInt, we must ensure we work with the
same type while calculating the total idle time, and only by the end
when we need to fetch idle ratio / utilization percentage, we convert
back to Number to do the final math & return.

If we don't use the same data type we run into data type errors trying
to calculate a total between BigInt and other types.