Integrate utilization tracker with Express/Fastify
This will enable tracking of active requests and idle time within those
frameworks via our middleware/plugin respectively, and include the `up`
metric with each WebMetricsCollector when collecting metrics for
reporting, so that they get sent back to our API.

I chose to not create a global utilization tracker but pass one to the
`WebMetricsCollector` since both the collector and tracker were being
initialized in the same place, and only the collector needs to reference
the tracker to collect metrics (other than the actual request/response
logic to track active requests), so it becomes a "global" in the sense
of the middleware/plugin it's currently in, not in the sense of a
Judoscale library global.

It should effectively work the same since we shouldn't have the same app
running both express and fastify anyway, and even if they do, these
should be scoped separately with a single reporter but different web
metrics collectors & utilization trackers. (but having two web processes
can be problematic anyway for other reasons, since we can't
differentiate their web metrics if they're sent to the same endpoint.)